requirements.txt
Django==1.10.1
djangorestframework==3.4.6
coverage==4.2
flake8<3.0.0
mkdocs==0.15.3
